Added Routes, Moved Logic, Added Logos and Icons
Added Browser Routers in the App.jsx, and moves Josh's code from App.jsx to HomePage.jsx.
Also added the Logo SVGs to be used throughout the site, along with icons for the login section.
Added the typekit from Adobe for the font Museo Sans and that's in the App.css